Moves sections folder into ui folder
