+++ components/ui/sections/LegacySection.vue
