+++ pages/education/index.vue
@@ -99,7 +99,7 @@ import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/logic/QiskitPage.vue'
 import InnerNavigation from '~/components/ui/InnerNavigation.vue'
 import CommunityHeader from '~/components/ui/CommunityHeader.vue'
-import PageSection from '~/components/sections/PageSection.vue'
+import PageSection from '~/components/ui/sections/PageSection.vue'
 import TextbookPreview from '~/components/education/TextbookPreview.vue'
 import Cta from '~/components/ui/Cta.vue'
 // @ts-ignore: Cannot find module
