+++ pages/experiments/_experiment.vue
@@ -23,7 +23,7 @@ import { Component } from 'vue-property-decorator'
 import { Context } from '@nuxt/types'
 import QiskitPage from '~/components/logic/QiskitPage.vue'
 import ExperimentHeader from '~/components/experiments/ExperimentHeader.vue'
-import PageSection from '~/components/sections/PageSection.vue'
+import PageSection from '~/components/ui/sections/PageSection.vue'
 import MdContent from '~/components/ui/MdContent.vue'
 
 @Component({
