+++ pages/ibmqaccount.vue
@@ -58,7 +58,7 @@ print(backend, backend.status().pending_jobs)</SyntaxHighlight>
 import { Component } from 'vue-property-decorator'
 import QiskitElementPage from '~/components/logic/QiskitElementPage.vue'
 import LegacyPresentation from '~/components/elements/LegacyPresentation.vue'
-import LegacySection from '~/components/sections/LegacySection.vue'
+import LegacySection from '~/components/ui/sections/LegacySection.vue'
 import SoftwareStack from '~/components/ui/SoftwareStack.vue'
 import SyntaxHighlight from '~/components/ui/SyntaxHighlight.vue'
 import Cta from '~/components/ui/Cta.vue'
